1.2.2
I add the pre progressing module in the system.  U can use the slider
to edit the Image. and it will save automatically in the computer. I
haven’t wrote the two buttons’ functions.And there is  a problem that
you after you do the pre progressing in label_pre,the image can’t show
on the label2, u have to open the saved image again.